Wanted to get away from the constructor parameter in DaoTransactionsFluxImpl
